fixed bootloader so it can actually boot from my disk

+++ drivers/pci.h
@@ -1,5 +1,6 @@
 #include <stdint.h>
-uint16_t pciReadWord(uint8_t bus ,uint8_t device, uint8_t function, uint8_t reg);
+#include "resolvepci.h"
+uint32_t pciReadWord(uint8_t bus ,uint8_t device, uint8_t function, uint8_t reg);
 uint16_t pcicheck(uint8_t bus,uint8_t device);
